+++ packages/core/test/bundling/forms_reactive/bundle.golden_symbols.json
@@ -1116,9 +1116,6 @@
     "name": "isFormControl"
   },
   {
-    "name": "isFormGroup"
-  },
-  {
     "name": "isForwardRef"
   },
   {
