+++ packages/core/test/bundling/forms_template_driven/bundle.golden_symbols.json
@@ -1077,9 +1077,6 @@
     "name": "isDirectiveHost"
   },
   {
-    "name": "isFormGroup"
-  },
-  {
     "name": "isForwardRef"
   },
   {
