+++ packages/forms/src/directives/form_interface.ts
@@ -6,7 +6,8 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {FormControl, FormGroup} from '../model';
+import {FormControl} from '../model/form_control';
+import {FormGroup} from '../model/form_group';
 
 import {AbstractFormGroupDirective} from './abstract_form_group_directive';
 import {NgControl} from './ng_control';
